pyproject.toml
[project]
name = "floor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.1",
    "pygame-ce>=2.5.5",
]
